Added AGDS data loading.

+++ src/agds/ValueNode.java
@@ -5,10 +5,10 @@ import java.util.List;
 /**
  * Created by biela.arek@gmail.com (Arek Biela) on 14.03.2016.
  */
-public class ValueNode extends Node<Float> {
+public class ValueNode extends Node<Double> {
 
 
-    public ValueNode(Float value, RecordNode recordNode) {
+    public ValueNode(Double value, RecordNode recordNode) {
         super(value);
         addNode(recordNode);
     }
